tukorea_notice_bot/requirements.txt
requests
beautifulsoup4
pytest
python-dotenv
